requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Security==1.7.4
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
blinker==1.3
boto==2.33.0
celery==3.1.16
ecdsa==0.11
itsdangerous==0.24
kombu==3.0.23
paramiko==1.15.1
passlib==1.6.2
pycrypto==2.6.1
python-dateutil==2.2
pytz==2014.7
six==1.8.0
wsgiref==0.1.2
